Add note to performance section about memory allocator lock contention.
I tested littlesat compiled against musl libc (first natively on an Adélie
machine and then with a cross toolchain on the Fedora machine), and it
is *significantly* slower due to lock contention in musl's current malloc
implementation. Apparently this is a known issue, which mallocng is intended
to address.